docs(vuejs): Rename result-label to result-text-field
https://coveord.atlassian.net/browse/KIT-1474
