+++ packages/samples/vuejs/src/templates/result-template.html
@@ -25,13 +25,13 @@
   <atomic-result-section-bottom-metadata>
     <atomic-result-fields-list>
       <atomic-field-condition class="field" if-defined="cat_platform">
-        <result-label label="Platform" field="cat_platform"></result-label>
+        <result-text-field label="Platform" field="cat_platform"></result-text-field>
       </atomic-field-condition>
       <atomic-field-condition class="field" if-defined="cat_condition">
-        <result-label label="Condition" field="cat_condition"></result-label>
+        <result-text-field label="Condition" field="cat_condition"></result-text-field>
       </atomic-field-condition>
       <atomic-field-condition class="field" if-defined="cat_categories">
-        <result-label label="Tags" field="cat_categories"></result-label>
+        <result-text-field label="Tags" field="cat_categories"></result-text-field>
       </atomic-field-condition>
     </atomic-result-fields-list>
   </atomic-result-section-bottom-metadata>
